Move cutlist types to a dedicated module
This leaves the root module with only the logic for interacting with
Fusion, which should be easier to follow. It also removes the use of a
global tolerance value in favor of passing the tolerance directly where
needed.
